pyproject.toml
[tool.poetry]
name = "async-tg-bot"
version = "0.1.0"
description = "The bot was implemented as the final project in the AAA Advanced Python course."
authors = ["yavnolib <yarik.nikk@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.7"
numpy = "^1.26.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
